// gomoku-web-ang/src/app/shared/enums/game-status.enum.ts
export enum GameStatusEnum {
  WAITING = "WAITING",
  CREATED = "CREATED",
  IN_PROGRESS = "IN_PROGRESS",
  BLACK_TO_MOVE = "BLACK_TO_MOVE",
  WHITE_TO_MOVE = "WHITE_TO_MOVE",
  BLACK_WINS = "BLACK_WINS",
  WHITE_WINS = "WHITE_WINS",
  DRAW = "DRAW",
  CANCELLED = "CANCELLED",
  UNKNOWN = "UNKNOWN"
}